Reverting to the revision with DAO classes and all tests passing.
While upgrading to svn revision 7077, I committed harakiri by trying to merge
changes by rsync from the svn repo folder to get the changes from 7088. This
was done to save time but by breaking from the routine, considerable number
of changes that used DAO classes were overwritten and there were no "conflict"
files that could be edited.

So, going back to the version which used the DAO classes and had all the tests
passing. Getting back to the latest version will be lot of work but more of a
known territory. Trying to merge the latest changes bringing up the old modifications
to make use of DAO classes will be troublesome to say the least.