Plugins group tests passing with mysql and postgresql
Tests in the Plugin group are passing with both mysql and postgresql.
Added the fetchAll and insertAll functions to the dao classes; these will be
used for setting up and tearing down the database in test cases. These
functions take into account the the differences in handling the binary
columns.
LoggerApiCall, LoggerError, LoggerException, LoggerMessage and Session dao classes
have been added so that the setup and teardown do not throw exceptions due to
the lack of dao classes for the logger_api_call, logger_error, logger_exception,
logger_message and session tables.

Added a function to the generic dao classes that check if the locking is
enabled.

Modified the deleteAll of the postgresql generic dao; this change makes it easy
to work with all kinds of select queries.

Added the "order by" clause to the site dao "getAllGroups" methods just to get
the ordering right so that the test passes.