Integration test TranstionsTest.php with postgres.
Integration/TransitionsTest.php is passing with both postgres and mysql.

core/ArchiveProcessing/Day.php
Added the $addGroupBy parameter to the queryActionsByDimension and
queryVisitsByDimension functions. This is to make sure that all of the
non-aggregate columns in the select clause are included in the group-by
clause.

core/Db/Helper/Pgsql/RankingQuery.php
Replaced the hard-coded -1 with $this->otherLabelValue.

plugins/Transitions/Transitions.php
Added the $addGroupBy columns to include all the non-aggregate columns
of select clause in the group by clause.
Modified the 'othersLabelValue' of RankingQuery to use '-1' instead of
'others'. Postgresql complained about non-integer values for some of
the columns, url_prefix.

tests/integration/expected/test_Transitions__Transitions.getFullReport_day.xml
Replaced 'Others' with '-1' to take into account the changes made in
the plugins/Transitions/Transitions.php